feat: refactor classes to optimize resources
One crazy big commit to enhance the memory usage of the classes. They
are now frozen dataclasses which of course implement by default __eq__
and __hash__ dunder methods to have a fingerprint of the instance of the
examined cipher.

Moreover, the README.md file has the correct size for the logo and the
essential information for any user or developer.

BREAKING CHANGE: classes to frozen dataclasses